==> notebooks/1. Data Cleaning.ipynb <==
import numpy as np
import pandas as pd

from bs4 import BeautifulSoup
import requests

data = pd.read_csv('../input_data/data.csv', sep='\t')
data.head()

data.info()

data.dtypes

data.shape

data.drop(['race', 'age', 'engnat', 'hand', 'source', 'country'], axis=1, inplace=True)

data.head()

### 1. Missing Values

data.isna().sum()

# Column: questions

columns = list(data.columns)[7:]

# Set of possible answers to the questions
ans = set(val 
          for col in columns 
          for val in data[col].unique())

print(ans)

data['E1'].unique()

filt = data['E1'] == 0
data[filt]

filt = data['E1'] == 0
idx = data.loc[filt].index.values[0]

data.drop(index=idx, inplace=True)
data.reset_index(inplace=True, drop=True)

data.shape

columns = list(data.columns)[7:]

# Set of possible answers to the questions after dropping the row
ans = set(val 
          for col in columns 
          for val in data[col].unique())

print(ans)

### Column: gender

gender = {0: 'Unknown',
          1: 'Male', 
          2: 'Female',
          3: 'Other'}

data['gender'].replace(gender, inplace=True)

filt = data['gender'] == 'Unknown'
n = data.loc[filt, 'gender'].count()
m = data.shape[0]
p = 100*n/m
percentage = round(p,2)

print('percentage:', str(percentage) + '%')

filt = data['gender'] == 'Other'
n = data.loc[filt, 'gender'].count()
m = data.shape[0]
p = 100*n/m
percentage = round(p,2)

print('percentage:', str(percentage) + '%')

filt = data['gender'] == 'Unknown'
data[filt]

idx = list(data.loc[filt].index.values)
data.drop(index=idx, inplace=True)

# Dropped 24 rows
data.shape

filt = data['gender'] == 'Other'
data[filt]

idx = list(data.loc[filt].index.values)
data.drop(index=idx, inplace=True)

data['gender'].unique()

### Data Output

data.to_csv('../output_data/data_cleaned.csv', index=False)
