==> notebooks/2. Table DB.ipynb <==
import numpy as np
import pandas as pd

data = pd.read_csv('../output_data/data_cleaned.csv')
data.head()

data.columns

questions = {'short': ['E1', 'E2', 'E3', 'E4', 'E5', 'E6', 'E7', 'E8', 'E9', 'E10',
                       'A1', 'A2', 'A3', 'A4', 'A5', 'A6', 'A7', 'A8', 'A9', 'A10', 
                       'C1', 'C2', 'C3', 'C4', 'C5', 'C6', 'C7', 'C8', 'C9', 'C10',
                       'N1', 'N2', 'N3', 'N4', 'N5', 'N6', 'N7', 'N8', 'N9', 'N10',
                       'O1', 'O2', 'O3', 'O4', 'O5', 'O6', 'O7', 'O8', 'O9', 'O10'],
             'long': ["I am the life of the party.", 
                      "I don't talk a lot.", 
                      "I feel comfortable around people.", 
                      "I keep in the background.", 
                      "I start conversations.", "I have little to say.", 
                      "I talk to a lot of different people at parties.", 
                      "I don't like to draw attention to myself.", 
                      "I don't mind being the center of attention.", 
                      "I am quiet around strangers.", 
                      
                      "I get stressed out easily.", 
                      "I am relaxed most of the time.", 
                      "I worry about things.", 
                      "I seldom feel blue.", 
                      "I am easily disturbed.", 
                      "I get upset easily.", 
                      "I change my mood a lot.", 
                      "I have frequent mood swings.", 
                      "I get irritated easily.", 
                      "I often feel blue.",                        
                      
                      "I feel little concern for others.", 
                      "I am interested in people.", 
                      "I insult people.", 
                      "I sympathize with others' feelings.", 
                      "I am not interested in other people's problems.", 
                      "I have a soft heart.", 
                      "I am not really interested in others.",  
                      "I take time out for others.",
                      "I feel others' emotions.",
                      "I make people feel at ease.",  
                      
                      "I am always prepared.", 
                      "I leave my belongings around.", 
                      "I pay attention to details.", 
                      "I make a mess of things.", 
                      "I get chores done right away.", 
                      "I often forget to put things back in their proper place.", 
                      "I like order.", 
                      "I shirk my duties.", 
                      "I follow a schedule.", 
                      "I am exacting in my work.",                      

                      "I have a rich vocabulary.", 
                      "I have difficulty understanding abstract ideas.", 
                      "I have a vivid imagination.", 
                      "I am not interested in abstract ideas.", 
                      "I have excellent ideas.", 
                      "I do not have a good imagination.", 
                      "I am quick to understand things.", 
                      "I use difficult words.", 
                      "I spend time reflecting on things.", 
                      "I am full of ideas."],
             'key': [+1, -1, +1, -1, +1, -1, +1, -1, +1, -1,
                     -1, +1, -1, +1, -1, -1, -1, -1, -1, -1, 
                     -1, +1, -1, +1, -1, +1, -1, +1, +1, +1, 
                     +1, -1, +1, -1, +1, -1, +1, -1, +1, +1, 
                     +1, -1, +1, -1, +1, -1, +1, +1, +1, +1], 
             'trait': ['Extraversion']*10 + \
                      ['Neuroticism']*10 + \
                      ['Agreeableness']*10 + \
                      ['Conscientiousness']*10 + \
                      ['Openness']*10}

data = pd.DataFrame(questions)

data.head(50)

data.to_csv('../output_data/questions_table.csv', index=False)
