==> notebooks/3. Completed data.ipynb <==
import numpy as np
import pandas as pd

table = pd.read_csv('../output_data/questions_table.csv')
table.head()

data = pd.read_csv('../output_data/data_cleaned.csv')
data.head()

filt = list(data.columns)[1:]
aux = data[filt]

aux = aux*list(table['key'])
aux.head()

filt = list(aux.columns)[30:40]
aux[filt]

aux.columns

traits = ['Extraversion', 'Neuroticism', 'Agreeableness', 'Conscientiousness', 'Openness']
print(traits)

filt = list(aux.columns)
seq = lambda x: filt[10*x:10*x + 10]

for i in range(len(traits)):
    data[traits[i]] = aux[seq(i)].sum(axis=1)

data.head()

data.to_csv('../output_data/data_complete.csv', index=False)
