==> notebooks/4. Exploratory Data Analysis - Gender.ipynb <==
import numpy as np
import pandas as pd

import plotly.express as px
import plotly.offline as pyo
import plotly.graph_objects as go
import plotly.figure_factory as ff

df = pd.read_csv('../output_data/data_complete.csv')
df.head()

# Gender Analysis

# Plot using a bar chart

gender_count = df['gender'].value_counts()

x = gender_count.index.values.tolist()
y = gender_count.values.tolist()

trace = go.Bar(x=x, y=y)

data = [trace]
layout = go.Layout(title='Gender',
                   xaxis={'title': 'Gender'},
                   yaxis={'title': 'Count'})

fig = go.Figure(data=data, layout=layout)
pyo.plot(fig, filename='gender.html')

# Traits Analysis

filt_m = df['gender'] == 'Male'
filt_f = df['gender'] == 'Female'

### Extraversion

male = df.loc[filt_m, 'Extraversion']
female = df.loc[filt_f, 'Extraversion']

data = [go.Box(y=male, name='Male', boxpoints='outliers', jitter=0.3, pointpos=0), 
        go.Box(y=female, name='Female', boxpoints='outliers', jitter=0.3, pointpos=0)]

layout = go.Layout(title='Extraversion scores by gender',
                   xaxis={'title': 'Gender'},
                   yaxis={'title': 'Score'})

fig = go.Figure(data=data, layout=layout)
pyo.plot(fig, filename='Extraversion.html')

hist_data = [male, female]
group_labels = ['Male', 'Female']

fig = ff.create_distplot(hist_data, 
                         group_labels, 
                         bin_size=[1,1])

pyo.plot(fig, filename='distplot1.html')

# Hypothesis testing

# H1: The two populations have different mean scores for Agreeableness

from scipy.stats import ttest_ind

(statistic, pvalue) = ttest_ind(male, female, equal_var=False)
print(pvalue)

# Neuroticism

male = df.loc[filt_m, 'Neuroticism']
female = df.loc[filt_f, 'Neuroticism']

data = [go.Box(y=male, name='Male', boxpoints='outliers', jitter=0.3, pointpos=0), 
        go.Box(y=female, name='Female', boxpoints='outliers', jitter=0.3, pointpos=0)]

layout = go.Layout(title='Neuroticism',
                   xaxis={'title': 'Gender'},
                   yaxis={'title': 'Score'})

fig = go.Figure(data=data, layout=layout)
pyo.plot(fig, filename='Neuroticism.html')

hist_data = [male, female]
group_labels = ['Male', 'Female']

fig = ff.create_distplot(hist_data, 
                         group_labels, 
                         bin_size=[1,1])

pyo.plot(fig, filename='distplot1.html')

# Hypothesis testing

# H1: The two populations have different mean scores for Agreeableness

from scipy.stats import ttest_ind

(statistic, pvalue) = ttest_ind(male, female, equal_var=False)
print(pvalue)

# Agreeableness

male = df.loc[filt_m, 'Agreeableness']
female = df.loc[filt_f, 'Agreeableness']

data = [go.Box(y=male, name='Male', boxpoints='outliers', jitter=0.3, pointpos=0), 
        go.Box(y=female, name='Female', boxpoints='outliers', jitter=0.3, pointpos=0)]

layout = go.Layout(title='Agreeableness',
                   xaxis={'title': 'Gender'},
                   yaxis={'title': 'Score'})

fig = go.Figure(data=data, layout=layout)
pyo.plot(fig, filename='Agreeableness.html')

male = df.loc[filt_m, 'Agreeableness']
female = df.loc[filt_f, 'Agreeableness']

hist_data = [male, female]
group_labels = ['Male', 'Female']

fig = ff.create_distplot(hist_data, 
                         group_labels, 
                         bin_size=[1,1])

pyo.plot(fig, filename='distplot1.html')

# Hypothesis testing

# H1: The two populations have different mean scores for Agreeableness

from scipy.stats import ttest_ind

(statistic, pvalue) = ttest_ind(male, female, equal_var=False)
print(pvalue)

# Conscientiousness

male = df.loc[filt_m, 'Conscientiousness']
female = df.loc[filt_f, 'Conscientiousness']

data = [go.Box(y=male, name='Male', boxpoints='outliers', jitter=0.3, pointpos=0), 
        go.Box(y=female, name='Female', boxpoints='outliers', jitter=0.3, pointpos=0)]

layout = go.Layout(title='Conscientiousness',
                   xaxis={'title': 'Gender'},
                   yaxis={'title': 'Score'})

fig = go.Figure(data=data, layout=layout)
pyo.plot(fig, filename='Conscientiousness.html')

hist_data = [male, female]
group_labels = ['Male', 'Female']

fig = ff.create_distplot(hist_data, 
                         group_labels, 
                         bin_size=[1,1])

pyo.plot(fig, filename='distplot1.html')

# Hypothesis testing

# H1: The two populations have different mean scores for Agreeableness

from scipy.stats import ttest_ind

(statistic, pvalue) = ttest_ind(male, female, equal_var=False)
print(pvalue)

# Openness

male = df.loc[filt_m, 'Openness']
female = df.loc[filt_f, 'Openness']

data = [go.Box(y=male, name='Male', boxpoints='outliers', jitter=0.3, pointpos=0), 
        go.Box(y=female, name='Female', boxpoints='outliers', jitter=0.3, pointpos=0)]

layout = go.Layout(title='Openness',
                   xaxis={'title': 'Gender'},
                   yaxis={'title': 'Score'})

fig = go.Figure(data=data, layout=layout)
pyo.plot(fig, filename='Openness.html')

male = df.loc[filt_m, 'Openness']
female = df.loc[filt_f, 'Openness']

hist_data = [male, female]
group_labels = ['Male', 'Female']

fig = ff.create_distplot(hist_data, 
                         group_labels, 
                         bin_size=[1,1])

pyo.plot(fig, filename='distplot1.html')

# Hypothesis testing

# H1: The two populations have different mean scores for Agreeableness

from scipy.stats import ttest_ind

(statistic, pvalue) = ttest_ind(male, female, equal_var=False)
print(pvalue)
